feat: protected routes and jwt

+++ backend/controllers/studentController.js
@@ -4,7 +4,16 @@ const Slot = require('../models/Slot');
 // GET /api/students
 exports.getAllStudents = async (req, res) => {
   try {
-    const students = await User.find({ role: 'Student' });
+    if 
+    (
+      req.user.role !== "Admin" &&
+      req.user.role !== "SysAdmin"
+    ) {
+      const users = await User.find({ role: { $in: ['Student', 'Tutor'] }, _id: req.user.id });
+      return res.status(201).json(users);
+    }
+
+    const students = await User.find({ role: { $in: ['Student', 'Tutor'] } });
     res.json(students);
   } catch (err) {
     console.error(err);
@@ -15,7 +24,22 @@ exports.getAllStudents = async (req, res) => {
 // GET /api/students/:studentId
 exports.getStudentById = async (req, res) => {
   try {
-    const student = await User.findOne({ _id: req.params.studentId, role: 'Student' });
+    if (
+      req.user.role !== "Student" &&
+      req.user.role !== "Admin" &&
+      req.user.role !== "SysAdmin"
+    ) {
+      return res.status(403).json({ message: "Access denied" });
+    }
+
+    if (
+      req.user.role !== "SysAdmin" &&
+      req.user.role !== "Admin" &&
+      req.params.studentId !== req.user.id
+    ) {
+      return res.status(403).json({ message: "Access denied" });
+    }
+    const student = await User.findOne({ _id: req.params.studentId, role: { $in: ['Student', 'Tutor'] } });
     if (!student) return res.status(404).json({ message: 'Student not found' });
     res.json(student);
   } catch (err) {
@@ -27,6 +51,14 @@ exports.getStudentById = async (req, res) => {
 // GET /api/students/:studentId/bookings
 exports.getStudentBookings = async (req, res) => {
   try {
+    if (
+      req.user.role !== "SysAdmin" &&
+      req.user.role !== "Admin" &&
+      req.params.studentId !== req.user.id
+    ) {
+      return res.status(403).json({ message: "Access denied" });
+    }
+
     const bookings = await Slot.find({ studentId: req.params.studentId }).populate('tutorId');
     res.json(bookings);
   } catch (err) {
@@ -38,9 +70,15 @@ exports.getStudentBookings = async (req, res) => {
 // PATCH /api/students/:studentId
 exports.updateStudent = async (req, res) => {
   try {
+    if (
+      req.user.role !== "SysAdmin" &&
+      req.params.studentId !== req.user.id
+    ) {
+      return res.status(403).json({ message: "Access denied" });
+    }
     const updates = req.body;
     const student = await User.findOneAndUpdate(
-      { _id: req.params.studentId, role: 'Student' },
+      { _id: req.params.studentId, role: { $in: ['Student', 'Tutor'] } },
       updates,
       { new: true }
     );
@@ -55,9 +93,14 @@ exports.updateStudent = async (req, res) => {
 // DELETE /api/students/:studentId
 exports.deleteStudent = async (req, res) => {
   try {
+    if (
+      req.user.role !== "SysAdmin"
+    ) {
+      return res.status(403).json({ message: "Access denied" });
+    }
     const result = await User.findOneAndDelete({
       _id: req.params.studentId,
-      role: 'Student'
+      role: { $in: ['Student', 'Tutor'] }
     });
     if (!result) return res.status(404).json({ message: 'Student not found' });
     res.json({ message: 'Student deleted successfully' });
